Added functional maps, species id in DB
The prevous vrsion of distribution maps framed the skeleton for the current version. Here the data is catched and formed in an effective manner. If the distribution data is unavailable, the the maps section is dynamically removed. The page uses pure JavaScript to perform all the required operations. Also added the Taxonomic ID in the Database files to better catch the spatial data.